PODS-9761 - Refactored package name
